requirements.txt
Django==2.2.9
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.11.0
mysqlclient==2.0.3
pytz==2020.5
sqlparse==0.4.1
